Refactor: 코드 간격 포메팅 실행
