+++ server/src/main/java/com/ryc/api/v2/announcement/domain/enums/AnnouncementStatus.java
@@ -1,7 +1,7 @@
 package com.ryc.api.v2.announcement.domain.enums;
 
 public enum AnnouncementStatus {
-    UPCOMING,
-    RECRUITING,
-    CLOSED
+  UPCOMING,
+  RECRUITING,
+  CLOSED
 }
