+++ server/src/main/java/com/ryc/api/v2/announcement/presentation/dto/response/AnnouncementGetAllResponse.java
@@ -3,5 +3,4 @@ package com.ryc.api.v2.announcement.presentation.dto.response;
 import lombok.Builder;
 
 @Builder
-public record AnnouncementGetAllResponse() {
-}
+public record AnnouncementGetAllResponse() {}
