+++ server/src/main/java/com/ryc/api/v2/club/presentation/dto/response/AllClubGetResponse.java
@@ -2,10 +2,10 @@ package com.ryc.api.v2.club.presentation.dto.response;
 
 import java.util.List;
 
-import com.ryc.api.v2.announcement.domain.enums.AnnouncementStatus;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import com.ryc.api.v2.announcement.domain.enums.AnnouncementStatus;
 import com.ryc.api.v2.club.domain.Category;
 import com.ryc.api.v2.club.domain.ClubTag;
 
@@ -26,5 +26,4 @@ public record AllClubGetResponse(
     @Schema(description = "동아리 태그 리스트") @NotNull(message = "club tag shouldn't be null")
         List<ClubTag> clubTags,
     @Schema(description = "동아리 공고 상태") @NotNull(message = "announcement status shouldn't be null")
-    AnnouncementStatus announcementStatus
-    ) {}
+        AnnouncementStatus announcementStatus) {}
