+++ server/src/test/java/com/ryc/api/v2/club/business/ClubServiceTest.java
@@ -1,4 +1,4 @@
-package com.ryc.api.v2.club.service;
+package com.ryc.api.v2.club.business;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
@@ -8,7 +8,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import com.ryc.api.v2.club.business.ClubService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
